Change how attribute nodes are stored and accessed
We used dictionaries for this earlier but this is downsides and no significant benefits. The number of attributes to be expected in a single element node is expected to be fairly low. Low enough that a linear search will likely be faster than a hash table lookup. Secondly hash tables don't preserve order, which regular arrays do.